refactored styling by adding CSS variables to maintain consistency, and changed folder structure a bit for clarity
