--- app/components/Content/index.js
@@ -1,1 +1,0 @@
-export { Content, links as contentLinks } from './Content';
