--- app/components/Tabs/Tab/Tab.css
@@ -1,70 +1,0 @@
-button[role='tab'] {
-  border: none;
-  text-align: center;
-  background-color: transparent;
-  font-family: 'League Spartan';
-  font-size: var(--9px);
-  line-height: 10px;
-  display: inline-block;
-  letter-spacing: 1.92857px;
-  text-transform: uppercase;
-  padding: 0;
-  color: var(--light);
-  position: relative;
-  cursor: pointer;
-
-  height: 40px;
-}
-
-button[role='tab'][aria-selected='true'] {
-  color: var(--light);
-}
-
-button[role='tab'][aria-selected='true']::after {
-  content: '';
-  position: absolute;
-  display: block;
-  height: 3px;
-  background-color: var(--backgroundColor);
-  width: 100%;
-  bottom: 0;
-  margin: 0;
-  border-radius: 0;
-}
-
-button[role='tab'] span {
-  display: none;
-}
-
-@media (min-width: 768px) {
-  button[role='tab'] {
-    border: 1px solid var(--grey-100);
-    text-align: start;
-    padding: 0 1.2rem;
-    width: 100%;
-  }
-
-  button[role='tab']:hover {
-    background-color: var(--grey-200);
-    border: 1px solid transparent;
-  }
-
-  button[role='tab'][aria-selected='true'] {
-    background-color: var(--backgroundColor);
-    border: 1px solid transparent;
-  }
-
-  button[role='tab'][aria-selected='true']::after {
-    display: none;
-  }
-
-  button[role='tab'] span {
-    display: inline;
-    color: var(--grey-100);
-    margin-inline-end: 1rem;
-  }
-
-  button[role='tab'][aria-selected='true'] span {
-    color: var(--light);
-  }
-}
