--- app/components/Tabs/Tab/index.js
@@ -1,1 +1,0 @@
-export { Tab, links as tabLinks } from './Tab';
