+++ app/components/Tabs/TabContent/TabContent.css
@@ -6,10 +6,14 @@
 }
 
 .content .content-image-wrapper {
-  margin-block: 4rem;
   margin-inline: auto;
   max-width: 400px;
   position: relative;
+}
+
+.content .content-image {
+  margin-inline: auto;
+  margin-block: 2rem;
 }
 
 .content .content-image-magnified {
@@ -47,12 +51,17 @@
   .content {
     display: grid;
     grid-template-columns: repeat(2, 50%);
-    grid-template-rows: 450px 1fr;
+    grid-template-rows: var(--gridRow) 1fr;
+    align-items: center;
     padding: 0 48px;
   }
 
   .content .content-image-wrapper {
     grid-column: 1 / -1;
+  }
+
+  .content .content-image {
+    margin-block: 0;
   }
 
   .content h2 {
@@ -68,8 +77,8 @@
 
 @media (min-width: 1024px) {
   .content {
-    grid-template-columns: 60% 40%;
-    grid-template-rows: 450px 200px;
+    grid-template-columns: var(--gridColumn);
+    grid-template-rows: var(--gridRowDesktop) 200px;
     place-items: center;
     padding: 0;
   }
