+++ app/components/Tabs/TabContent/index.js
@@ -1,0 +1,1 @@
+export { TabContent, links as tabContentLinks } from './TabContent';
